Finished TranslateDictionary, Began work on new scanLine
